Replace Doc2 with Stream[D2].
Rather than use a linked data type, this approach uses Scala's Stream
type to achieve laziness in the tail. This optimizes the best() method
by ensuring we only do enough work in fits() to determine if we need
to wrap a line or not.

This change ends up making the code simpler in a few places.