Implement comparisons for Doc.
This algorithm needs some thought put into it, but as a first-pass it
might possibly be correct.

The basic idea is this:

 1. Only the right-hand side of a union needs to be considered, since
    if those are equal we can be sure that they flatten to the same
    thing.

 2. Within the right-hand side of a union, only the presence (or
    absence) of newlines makes a difference -- all other text can be
    considered "the same" whether it is inside or outside a union.

 3. For a given newline, we need to track how many right-hand-side
    unions it is contained within, to determine the exact conditions
    when it may (or may not) be flattened. If two Docs have newlines
    at a given "position", they need to have the same "union
    structure" at that point to be equal.

Using these ideas, we generate a stream of `Tok` values (which are
similar, but not identical to, `Doc2` values). We then consume the
streams for two documents, looking for any possible discrepencies.

The comparison should be relatively efficient (O(n)), though in cases
where text concatenations differ there will be some number of extra
strings created via `substring`.