First pass at polishing.
This method adds good Scaladoc comments to every method in Doc. It
also removes/renames some methods to try to give a slightly better
user-experience. It is relatively opinionated.

The companion object of Doc still needs some work, and we probably
want to split up Doc.scala, but I'm pushing this now to get feedback.