Remove Ordering[Doc], DocTree, and friends.
We were spinning out on this a bit -- we are pretty sure it is
solvable but probably requires enough research and effort that it
doesn't make sense to hold up our release for it.

For testing purposes, we test with several different interesting
widths to have reasonable confidence that things which should be equal
for all renderings really are so.