WIP: Add hard line break that is unmodifed by 'grouped'
Following the Wadler/Leijen pretty printers in Haskell,
this change adds `FlatAlt` and `Fail` to add hard line
breaks.

This change is to solicit some feedback.  Some of the auxiliary
cases are not totally satisfactory, but all the tests pass.
Unlike the previous version without `Fail`, the change
is lighter weight and more elegant.