Replace renderWideStream with renderCompactStream
In the presence of hard line breaks, renderWideStream seems
hard to justify as a stack-safe procedure; if flattening
fails, it's not satisfying to consider it as a wide stream,
as the document is just rendered as is. It may be possible
to give it some concrete semantics, e.g. hard lines ignore
indent, but otherwise try to maximally lay out lines, but
it's hardly canonical in the same way it was when there were
no hard line breaks.

It seems used mainly to quickly implement hashCode, so this
change simply makes a quick compact document for that purpose.