Clarify Union invariants
This change came about when I was trying to understand the
`Union` invariants.  It's mostly a no-op.  I changed some
wording that I thought could be made more clear, grouped
the tests for `Union` invariants together, and added a
couple new ones that correspond to the invariants described
in the comments.