Add genDocNoFill generator (#133)
When we add hard line breaks, `fill` will put arbitrary documents on
the lhs of `Union`.  This will break numerous desirable laws, such as
`Union(a + b, a + c) == a + Union(b, c)`; when there's a `Line` in
`a`, it's not possible to render the two sides as equal at all lengths
in general.

To prepare for the change, this review adds a generator for documents
with Unions only generated by `grouped`, which is well behaved wrt the
existing laws/tests.