Replace Stream[A] with LazyIterable[A].
We don't need many of the features Stream provides (e.g. memoization),
so we can replace it with a very simple implementation that extends
Iterable. We also don't need to ensure that combinators such as map
are lazy, because we build these things up lazily but then take them
apart eagerly in most cases.

I'm open to adding more features to LazyIterable if we _do_ want to
make it a production-ready collection type. For now it's the smallest
thing I thought we could replace Stream with.