Add zero-width text nodes and styled nodes. (#184)
* Add zero-width text nodes and styled nodes.

This change introduces a new concept: ZeroWidth(_). This is basically
a Text(s) node whose internal string doesn't count against the
"budget" that Paiges uses to render. The expectation is that this will
be used to support "invisible" pieces of text, such as ANSI control
characters. We do expose a zeroWidth() constructor, but this is a
power user feature and it's definitely possible to use this
incorrectly.

We also added a Styled(_, _) node, which applies a style to a Doc.
During rendering this will translate into zero-width nodes, but unlike
zero-width nodes, styles are composable. Currently styles are provided
for ANSI and XTerm control sequences.

* Improve test coverage and styling.

* Remove Doc.Styled nodes.

This removes some rendering complexity. We support the same public API
by implementing the .style() method with ZeroWidth nodes directly.

* Defensive check during zeroWidth constructor.

* Fix bad merge