Future-proof the previous version detection.
Previously we had a hand-coded (and hand-maintained) strategy that
would fail every time we released a new minor (or major) version of
Paiges. Now, we have a fairly reasonable heuristic that ensures we are
preserving compatibility across versions.

We aren't being as strict as we might want when bumping the minor or
major versions (we could be requiring compatibility with the
chronologically last release) but that would require a bigger change
and is outside the scope of this work for now.

(This isn't pressing because 0.x releases don't promise compatibility
across minor versions, so we don't need to solve this until 1.0.)